requirements.txt
alembic==1.8.1
anyio==3.6.1
async-asgi-testclient==1.4.11
asyncpg==0.26.0
attrs==22.1.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
databases==0.6.1
greenlet==1.1.3
h11==0.13.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
multidict==6.0.2
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn[standard]==0.18.3
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
